requirements.txt
pip==23.2.1
asgiref==3.7.2
Django==4.2.3
django-cockroachdb==4.2
django-cors-headers==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
mysqlclient==2.2.0
Pillow==10.0.0
psycopg2==2.9.7
PyJWT==2.8.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
